Added simple logger implementation
to add to the log, simply add these lines to your class:
ServerLogger logger = ServerLogger.getInstance();
logger.myLogger.info("Your message here");

info represents the priority of the message. can be severe, info,
finest, debug for example.

This implementation is not the best, as it follows the singleton
pattern. But for now, it does its job

Log output is saved to the Project path/temp.txt