Added an XMLTool in the main vooga util folder
The XMLTool was created to facilitate the creation and parsing of an
XML file. The code is still not complete and can accept many more
features and refactoring. The code has passed CheckStyle, but there are
further improvements pending with the code efficiency and comment
quality. The main functions were added to the XMLToolExample.java file.