Refactored the code to improve the versatility of this utility, improved comments and formatting.
This code was subject of accidental modification many times because of
the default variable RESOURCE_LOCATION = "/vooga/". Due to the presence
of many Pixmap classes in different branches and game genres, an error
in the import and in the compilation would lead a programmer to
accidentally modify the variable causing issues to the rest of the
contributors of this project.

This modification requires common agreement, since the implementation
of this code was slightly modified. Now the constructor and the method
setImage take the filePath including the filename, not the filename
only.